% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{base_net_ppi}
\alias{base_net_ppi}
\title{Base Network Processing For PPI Enrichment}
\usage{
base_net_ppi(i_net, vect, filt = "Edge")
}
\arguments{
\item{i_net}{the initial network object eg. basic_network}

\item{vect}{the edge attributes to tabulate}

\item{filt}{the edge attribute name to remove duplicates of default. Edge}
}
\description{
Processes a base network object for PPI enrichment tests. Takes the user 
supplied base network converts to a data frame. Removes duplicated edges.
Tabulates the probability of seeing an interaction of the types specified in 
`vect`.
}
\examples{
\dontrun{
data(basic_network)
base_net_ppi( 
 i_net = basic_network, 
 vect =c( 'Brain', 'Heart', 'Kidney', 'Liver',
    'Lung', 'Muscle', 'Thymus'
  ) 
)
}
}
