% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_network}
\alias{basic_network}
\title{Basic Network Object}
\format{
A list object
\describe{
  \item{igraph::V(basic_network)}{Vertex names}
  \item{igraph::E(basic_network)}{Network edges}
  \item{igraph::E(basic_network)$interaction}{Type of interaction}
  \item{igraph::E(basic_network)$Occurance}{How many Pathway Commons 
  databases record an interaction between to and from vertex}
  \item{igraph::E(basic_network)$UniqCol}{The Unique Code of the 
  entire interaction}
  \item{igraph::E(basic_network)$pathway}{The Pathway(s) in Pathway 
  commons this interaction is documented in as a list object}
  \item{igraph::E(basic_network)$EdgeRep}{How many interaction types 
  are there between to and from vertacies}
  \item{igraph::E(basic_network)$Edge}{The to:from name desigantion 
  of the edge}
  \item{igraph::E(basic_network)$SumOccurancel}{how many times is to
   vertex connected to from vertex across all pathway commons data bases and 
   interaction types}
 \item{igraph::E(basic_network)$Uniq_Brain}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to brain are denoted with a 1}
 \item{igraph::E(basic_network)$Brain}{Edges with a PPI in Skinnider 
   et.al 2021 that in brain are denoted with a 1}
 \item{igraph::E(basic_network)$Uniq_Heart}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to Heart are denoted with a 1}
 \item{igraph::E(basic_network)$Heart}{Edges with a PPI in Skinnider 
   et.al 2021 that in Heart are denoted with a 1}
 \item{igraph::E(basic_network)$Uniq_Kidney}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to Kidney are denoted with a 1}
 \item{igraph::E(basic_network)$Kidney}{Edges with a PPI in Skinnider 
   et.al 2021 that in Kidney are denoted with a 1}
 \item{igraph::E(basic_network)$Uniq_Liver}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to Liver are denoted with a 1}
 \item{igraph::E(basic_network)$Liver}{Edges with a PPI in Skinnider 
   et.al 2021 that in Liver are denoted with a 1}
 \item{igraph::E(basic_network)$Uniq_Lung}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to Lung are denoted with a 1}
 \item{igraph::E(basic_network)$Lung}{Edges with a PPI in Skinnider 
   et.al 2021 that in Lung are denoted with a 1}
 \item{igraph::E(basic_network)$Uniq_Muscle}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to Muscle are denoted with a 1}
 \item{igraph::E(basic_network)$Muscle}{Edges with a PPI in Skinnider 
   et.al 2021 that in Muscle are denoted with a 1}
 \item{igraph::E(basic_network)$Uniq_Thymus}{Edges with a PPI in Skinnider 
   et.al 2021 that is unique to Thymus are denoted with a 1}
 \item{igraph::E(basic_network)$Thymus}{Edges with a PPI in Skinnider 
   et.al 2021 that in Thymus are denoted with a 1}
 \item{igraph::E(basic_network)$all_nonbrain}{Edges with a PPI in Skinnider 
   et.al 2021 that is observed in any tissue except Braing are denoted with 
   a 1}
 \item{igraph::E(basic_network)$all_uniq_nonbrain}{Edges with a PPI in 
 Skinnider et.al 2021 that are observed uniquely in any tissue except Brain
 are denoted with a 1}
  ...
}
}
\usage{
basic_network
}
\description{
The basic Pathway Commons network. This is all SIF Pathway commons 
interactions loaded into a network. The CHEMI interactions are present. The
igraph network object consisting of 30,910 genes and 1,902,605 
 interactions. basic_network$graph is a list of networks. Each 
network is every interaction in Pathway Commons consisting of a specific type
of interaction.
}
\keyword{datasets}
