% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_network_interactions}
\alias{basic_network_interactions}
\title{Basic Network Interactions}
\format{
A list object
\describe{
  \item{basic_network_interactions}{A list of igraph network objects by interaction 
  type, all edge and vertex attributes are the same as basic_network$network}
  ...
}
}
\usage{
basic_network_interactions
}
\description{
This Resources is a list of networks. Each network is every interaction in 
Pathway Commons consisting of a specific type of interaction.
}
\keyword{datasets}
