% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_network_table}
\alias{basic_network_table}
\title{Basic Network Table}
\format{
A data table
\describe{
  \item{basic_network_table$to}{Origin Vertex}
  \item{basic_network_table$from}{Target Vertex}
  \item{basic_network_table$interaction}{Type of interaction}
  \item{basic_network_table$Occurance}{how many Pathway Commons databases record an
    interaction between to and from vertex}
  \item{basic_network_table$UniqCol}{The Unique Code of the entire interaction}
  \item{basic_network_table$pathway}{The Pathway(s) in Pathway commons this 
    interaction is documented in as a list object}
  \item{basic_network_table$EdgeRep}{How many interaction types are there between 
    to and from vertacies }
  \item{basic_network_table$Edge}{The to:from name desigantion of the edge}
  \item{basic_network_table$SumOccurancel}{how many times is to vertex connected to
    from vertex across all pathway commons data bases and interaction types}
  \item{basic_network_table$Uniq_Brain}{Edges with a PPI in Skinnider 
    et.al 2021 that is unique to brain are denoted with a 1}
  \item{basic_network_table$Brain}{Edges with a PPI in Skinnider 
    et.al 2021 that in brain are denoted with a 1}
  \item{basic_network_table$Uniq_Heart}{Edges with a PPI in Skinnider 
    et.al 2021 that is unique to Heart are denoted with a 1}
  \item{basic_network_table$Heart}{Edges with a PPI in Skinnider 
    et.al 2021 that in Heart are denoted with a 1}
  \item{basic_network_table$Uniq_Kidney}{Edges with a PPI in Skinnider 
    et.al 2021 that is unique to Kidney are denoted with a 1}
  \item{basic_network_table$Kidney}{Edges with a PPI in Skinnider 
    et.al 2021 that in Kidney are denoted with a 1}
  \item{basic_network_table$Uniq_Liver}{Edges with a PPI in Skinnider 
    et.al 2021 that is unique to Liver are denoted with a 1}
  \item{basic_network_table$Liver}{Edges with a PPI in Skinnider 
    et.al 2021 that in Liver are denoted with a 1}
  \item{basic_network_table$Uniq_Lung}{Edges with a PPI in Skinnider 
    et.al 2021 that is unique to Lung are denoted with a 1}
  \item{basic_network_table$Lung}{Edges with a PPI in Skinnider 
    et.al 2021 that in Lung are denoted with a 1}
  \item{basic_network_table$Uniq_Muscle}{Edges with a PPI in Skinnider 
    et.al 2021 that is unique to Muscle are denoted with a 1}
  \item{basic_network_table$Muscle}{Edges with a PPI in Skinnider 
     et.al 2021 that in Muscle are denoted with a 1}
   \item{basic_network_table$Uniq_Thymus}{Edges with a PPI in Skinnider 
     et.al 2021 that is unique to Thymus are denoted with a 1}
  \item{basic_network_table$Thymus}{Edges with a PPI in Skinnider 
    et.al 2021 that in Thymus are denoted with a 1}
  \item{basic_network_table$all_nonbrain}{Edges with a PPI in Skinnider 
    et.al 2021 that is observed in any tissue except Braing are denoted with 
    a 1}
  \item{basic_network_table$all_uniq_nonbrain}{Edges with a PPI in Skinnider 
    et.al 2021 that are observed uniquely in any tissue except Brain are 
    denoted with a 1}
  ...
}
}
\usage{
basic_network_table
}
\description{
The basic Pathway Commons table that comprises the foundation of the network
This is all SIF Pathway commons interactions loaded into a network. The 
CHEMI interactions are present. The igraph network object consisting of 
30,910 genes and 1,902,605 interactions.
}
\keyword{datasets}
