% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{derived_net_ppi}
\alias{derived_net_ppi}
\title{Derived Network Processing For PPI Enrichment}
\usage{
derived_net_ppi(d_net, vect, filt = "Edge")
}
\arguments{
\item{d_net}{a list of derived network object(s) eg. list(test_net_A)}

\item{vect}{the edge attributes to tabulate}

\item{filt}{the edge attribute name to remove duplicates of default. Edge}
}
\description{
Processes a a list object of derived network object(s) for PPI enrichment 
tests. Takes the user supplied base networks converts to a data frame.
Removes duplicated edges.
Tabulates the probability of seeing an interaction of the types specified in 
`vect`.
}
\examples{
\dontrun{
derived_net_ppi( 
 i_net = list(basic_network), 
 vect =c( 'Brain', 'Heart', 'Kidney', 'Liver',
    'Lung', 'Muscle', 'Thymus'
  ) 
)
}
}
