% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_limit}
\alias{find_limit}
\title{Finds the limit cutoff when target and sentinal paths are given}
\usage{
find_limit(s_path, t_path, weights, cores = 1)
}
\arguments{
\item{s_path}{the sentinel path object}

\item{t_path}{the target path object}

\item{weights}{OMICS/vertex scores as named vector}

\item{cores}{the cores to run the path calulation over. default = 1}
}
\value{
List object with named attributes of limit (the path score cutoff 
limit) and both t_verts and s_verts objects used as placeholders to determine
if the paths were scorable.
}
\description{
Limit is based on whichever value is greater between; the mean or median 
scores of the Target-traced-to-sentinal paths. If there are no target to
 sentinal paths the pairwise within target traces are used.
}
\examples{

example_path <- list()
example_path$res <- list(
   c('319', "49", "23", "86", "690", "238"),
   c('422', "899", "37", "240", "970", "28")
)
names(example_path$res[[1]]) <- c(
   "GeneA","GeneZ", "GeneAlpha",
   "GeneB", "GeneX", "GeneOmega"
)
names(example_path$res[[2]]) <- c(
   "Gene1","Gene2", "GeneUno",
   "Gene12", "Gene13", "GeneOcho"
)
example_path_b <-  example_path 
sampweights <- c(1.45, 2.45, 0.89, .003, 1.3, 2.1, 0.02, 0, 0, 0.2, 0.6, .70)
names(sampweights) <- c(
   "GeneA","GeneZ", "GeneAlpha",
   "GeneB", "GeneX", "GeneOmega",
   "Gene1","Gene2", "GeneUno",
   "Gene12", "Gene13", "GeneOcho"
)
find_limit(
   s_path = example_path,
   t_path =  example_path_b,
   weights = sampweights,
   cores = 2
)
}
