% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genelists}
\alias{genelists}
\title{Sample Genelists to Run Network Functions On}
\format{
An igraph network object
\describe{
  \item{targets}{target gene lists}
  \item{sentinals}{sentinal gene lists}
  ...
}
}
\usage{
genelists
}
\description{
Simple lists of genes with associated biological function. Target and 
sentinal gene lists contain the same 15 lists of genes stored as character
vectors in lists named targets and sentinals.
}
\keyword{datasets}
