% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_load}
\alias{list_load}
\title{Loads a Gene List From Synapse}
\usage{
list_load(file_path, network, is_syn = FALSE)
}
\arguments{
\item{file_path}{the igraph network to push to synapse eg. net}

\item{network}{igraph network object to use to filter the vertices From}

\item{is_syn}{is the list a synapse ID default=FALSE}
}
\value{
genes from the input present within the network
}
\description{
Loads a gene list. If is_syn is TRUE file_path is interperated as a synapse
ID. Otherwise file_path is interperated as a file path.
}
\examples{
all_goterms <- list(
   c('syn25185319', "APP_Metabolism", "APP Metabolism"),
   c('syn25185320', "Endolysosomal", "Endolysosomal")
)
list_load(
   all_goterms[[1]][1],
   network=<USER SPECIFIED igraph Network>,
   is_syn = TRUE
)
}
