% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_into_synapse}
\alias{log_into_synapse}
\title{Log into Synapse}
\usage{
log_into_synapse(usr = NULL, pass = NULL)
}
\arguments{
\item{usr}{A single value character vector of the users Synapse ID}

\item{pass}{A single value character vector of the users Synapse Password}
}
\value{
Synapse login object to use in list_load(), store_net(), and 
network_load()
}
\description{
Log into Synapse. Assumes credentials are stored.
User can specify UserID and Password as
usr and pass respectivly
}
\examples{
log_into_synapse()
log_into_synapse(
   usr = NULL,
   pass = NULL
 )
 
}
