% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_enrich}
\alias{plot_enrich}
\title{Plot PPI Enrichment}
\usage{
plot_enrich(enrich_obj, scaled = "NO")
}
\arguments{
\item{enrich_obj}{enrichment object from `ppi_enrichment`}

\item{scaled}{Scale the fold change by filter methods of each tissue 
default No. Values accepted; YES or NO}
}
\description{
Takes PPI enrichment object from `ppi_enrichment` and plots it with either
scaled or non-scaled values
}
\examples{
\dontrun{
data(basic_network)
base <- base_net_ppi( 
 i_net = basic_network, 
 vect =c( 'Brain', 'Heart', 'Kidney', 'Liver',
    'Lung', 'Muscle', 'Thymus'
  ) 
)
derived <- derived_net_ppi( 
 i_net = list(basic_network), 
 vect =c( 'Brain', 'Heart', 'Kidney', 'Liver',
    'Lung', 'Muscle', 'Thymus'
  ) 
)
ppi_enrichment(base=base, derived=derived)
plot_enrich(ppi_enrichment)
}
}
