% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ppi_enrichment}
\alias{ppi_enrichment}
\title{PPI Enrichment}
\usage{
ppi_enrichment(base, derived)
}
\arguments{
\item{base}{The Base Enrichment Network Stats from `base_net_ppi`}

\item{derived}{The Derived Enrichment Network Stats from `derived_net_ppi`}
}
\description{
Takes inputs of base network stats from `base_net_ppi` in the form of a list
and derived network stats from `derived_net_ppi` in the form of a list of 
lists.
}
\examples{
\dontrun{
data(basic_network)
base <- base_net_ppi( 
 i_net = basic_network, 
 vect =c( 'Brain', 'Heart', 'Kidney', 'Liver',
    'Lung', 'Muscle', 'Thymus'
  ) 
)
derived <- derived_net_ppi( 
 i_net = list(basic_network), 
 vect =c( 'Brain', 'Heart', 'Kidney', 'Liver',
    'Lung', 'Muscle', 'Thymus'
  ) 
)
ppi_enrichment(base=base, derived=derived)
}
}
