% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_network}
\alias{sample_network}
\title{Sample simple network}
\format{
An igraph network object
\describe{
  \item{verticies}{genes}
  \item{edges}{protien interactions}
  ...
}
}
\usage{
sample_network
}
\description{
A simple example network without weights
}
\keyword{datasets}
