% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{short_paths}
\alias{short_paths}
\title{Traces the shortest paths of a gene to a vector of target and a vector of 
sentinal genes}
\usage{
short_paths(tnet, target, targets, sentinals, cores = 1)
}
\arguments{
\item{tnet}{igraph network (Main entire network) eg. net/net_undirected/JS_net_undirected}

\item{target}{the from gene target eg Genes[1]}

\item{targets}{List of the total list of targets in the User set eg. Genes}

\item{sentinals}{List of the sentinal genes to trace to eg. Sentinal}

\item{cores}{the number of cores to use in paralellizable functions}
}
\value{
List object of Inter genes from target path traces and Sentinal genes
from sentinal gene traces
}
\description{
Traces the the shortest paths of target gene paiwise to the target gene list.
then traces the shortest path of the gene to a list of sentinal genes. This
function returns the list of genes in a path object which score over the an
emperically defined limit using the find_limit() function.
}
\examples{
data(slim_net, package = "igraphNetworkExpansion")
data(genelists, package = "igraphNetworkExpansion")

example_path <- short_paths(
    tnet = slim_net,
    target = genelists$targets$APP_Metabolism[1],
    targets = genelists$targets$APP_Metabolism,
    sentinals = genelists$sentinals$Immune_Response,
    cores = 1
)
}
