% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sif_boot}
\alias{sif_boot}
\title{Load and Add names to SIF formatted files}
\usage{
sif_boot(sif, sif_name, synap_import)
}
\arguments{
\item{sif}{a caracter vector value of a synID corresponding to a SIF file}

\item{sif_name}{the name of a variable corresponding to a vector value that
is a synID}

\item{synap_import}{is the reticulated imported synapse from 
log_into_synapse()$synapse eg. syn}
}
\value{
a dataframe object of the SIF file named the value of sif_name
}
\description{
The purpose of this function is to load SIF files direct form synapse
}
\examples{
\dontrun{
  syn <- igraphNetworkExpansion::log_into_synapse()
  sifs <- c('syn21914063')
  names(sifs) <- ('Detailed') 
  sif_file <- sif_boot( sifs, namse(sifs), synap_import )
}  
}
