% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sif_loader}
\alias{sif_loader}
\title{Load and Add namees to a list SIF formatted files}
\usage{
sif_loader(sifs, synap_import)
}
\arguments{
\item{sifs}{a caracter vector value of a synID corresponding to a SIF file}

\item{synap_import}{is the reticulated imported synapse from 
log_into_synapse()$synapse eg. syn}
}
\value{
a list of SIF dataframes named by the names of the input vector and 
        an igraph network object
}
\description{
The purpose of this function is to load multiple SIF files directly form synapse
}
\examples{
\dontrun{
  syn <- igraphNetworkExpansion::log_into_synapse()
  sifs <- c('syn21914063', 'syn21914056')
  names(sifs) <- ('Detailed', 'bind') 
  sif_file <- sif_boot( sifs, synap_import )
}  
}
