% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slim_net}
\alias{slim_net}
\title{Test and example network: implicit loading}
\format{
An igraph network object
\describe{
  \item{verticies}{genes}
  \item{edges}{protien interactions}
  \item{igraph::V(slim_net$name)}{The Vertex Gene Name}
  \item{igraph::V(slim_net$community)}{A Sample igraph Community designation 
  for the vertex}
  \item{igraph::V(slim_net$weight)}{A Sample Weight Value}
  \item{igraph::V(slim_net$RNA_EffectScore)}{A Sample RNA Phenotype 
  Effect Score}
  \item{igraph::V(slim_net$Pro_EffectScore)}{A Sample Protien Phenotype
  Effect Score}
  \item{igraph::V(slim_net$RNA_Cor_EffectScore)}{A Sample RNA co-expression 
  value}
  \item{igraph::V(slim_net$Pro_Cor_EffectScore)}{A Sample Protien co-expression 
  value}
  \item{igraph::E(slim_net$interaction)}{The Edge Interaction Type}
  \item{igraph::E(slim_net$UniqCol)}{The From_vert:To_vert-Interaction_type
  unique identifier}
  \item{igraph::E(slim_net$Edge)}{The From_vertex:To_vertex identifier}
  \item{igraph::E(slim_net$pathway)}{The Pathway Commons data base(s) where 
  this edge is found}
  \item{igraph::E(slim_net$EdgeRep)}{How many interaction types are there 
  between the two vertices across pathway commons databases}
  \item{igraph::E(slim_net$Occurance)}{How many Data Bases across pathway
  commons record an interaction of anytype between the two vertices}
  \item{igraph::E(slim_net$SumOccurence)}{Total amount of times these to 
  vertices interact accross all pathway commons database and interaction 
  types}
  ...
}
}
\usage{
slim_net
}
\description{
A Full Template network to use for testing and examples. All protein-protein
interactions in pathway commons for genes exressed in at least one tissue 
from AMP-AD transcriptomics data.
}
\keyword{datasets}
