% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{store_net}
\alias{store_net}
\title{Push Network to Synapse}
\usage{
store_net(
  network,
  net_filename,
  net_synname,
  p_id,
  folder,
  act_name,
  act_desc,
  code = NULL,
  repo = NULL,
  syn_used = NULL,
  subset = NULL,
  prov_object = NULL
)
}
\arguments{
\item{network}{the igraph network to push to synapse eg. net}

\item{net_filename}{the file name of the network without file extension}

\item{net_synname}{the desplay name of the network in synapse}

\item{p_id}{the parent synapse ID of the network destination}

\item{folder}{the name of the storage folder in the parent synapse ID to 
store the net}

\item{act_name}{the name of the syn activity object to}

\item{act_desc}{the description of the syn activity object to}

\item{code}{the path of the code which generated the network for the 
provenance (optional)}

\item{repo}{the repo which generated the network for the 
provenance (optional)}

\item{syn_used}{character vector of synIDs to seed the provenance (optional)}

\item{subset}{An vector of vertex names to filter the network for (optional) 
eg. test}

\item{prov_object}{A pre made github code provenance object or vector of 
objects}
}
\value{
a synapse entity of a .graphml subnetwork object stored in synapse
}
\description{
This function takes a network object and pushes it to synapse
}
\examples{

}
