% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_pull}
\alias{table_pull}
\title{Pull Synapse Table into Dataframe
The purpose. of this function is to pull info from a synapse table into a
dataframe.}
\usage{
table_pull(syn_id, feature_name, features, column_names, synap_import)
}
\arguments{
\item{syn_id}{the synapse ID of the ie 'syn25556478'}

\item{feature_name}{the column of the feature to treat as rows ie.'GeneName'}

\item{features}{the values of the feature to pull ie. names(igraph::V(net))}

\item{column_names}{the column values that. should be pulled for each 
feature-row ie c('ENSG', 'GeneName', 'OmicsScore', '
GeneticsScore', 'Logsdon')}

\item{synap_import}{is the reticulated imported synapse from 
log_into_synapse()$synapse eg. syn}
}
\value{
a dataframe object
}
\description{
Pull Synapse Table into Dataframe
The purpose. of this function is to pull info from a synapse table into a
dataframe.
}
\examples{
\dontrun{
syn <- log_into_synapse()

omics_scores <- table_pull( 
 syn_id ='syn25575156',
 feature_name = 'GeneName' ,
 features = names(igraph::V(net)),
 column_names = c('ENSG', 'GeneName', 'OmicsScore', 'GeneticsScore', 'Logsdon'),
 synap_import = syn$synapse
)
}
}
