% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trace_filter}
\alias{trace_filter}
\title{Process a path trace list}
\usage{
trace_filter(path_obj)
}
\arguments{
\item{path_obj}{path trace object from short_paths()}
}
\value{
a list of genes from the path trace to use for filtering the main
network
}
\description{
This function takes a path trace object from short_paths() and transforms it
into the list of genes to keep for the sub network generation.
}
\examples{
obj <- list(list( 
    Inter = c("GeneA","GeneZ", "GeneAlpha",
       "GeneB", "GeneX", "GeneOmega",
       "Gene1","Gene2", "GeneUno",
       "Gene12", "Gene13", "GeneOcho"),
    Sentinal = c("GeneEh","GeneZee", "GeneAlpha",
       "GeneB", "GeneXray", "GeneOmega",
       "Gene11","Gene21", "GeneUno",
       "Gene1", "Gene3", "GeneOcho")
))
trace_filter(obj)
}
