% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vertex_annotate}
\alias{vertex_annotate}
\title{Remove Duplicate Gene Entries
The purpose. of this function is to pull info from a synapse table into a
dataframe. Removal can be of all entries with a lower score in a given column
or entries which lack the desired ensg}
\usage{
vertex_annotate(df, ig_net, v_col, default_value = 0)
}
\arguments{
\item{df}{the data frame contain the vertex ids as renames}

\item{ig_net}{the network to annotate}

\item{v_col}{the column name of the feature become vertex attribute}

\item{default_value}{default vertex values if not present in df default = 0}
}
\value{
an annotated igraph network object
}
\description{
Remove Duplicate Gene Entries
The purpose. of this function is to pull info from a synapse table into a
dataframe. Removal can be of all entries with a lower score in a given column
or entries which lack the desired ensg
}
\examples{
\dontrun{
 data(slim_net, package = "igraphNetworkExpansion")
 syn <- log_into_synapse()
 omics_scores <- dplyr::left_join(
  table_pull(
    syn_id ='syn25575156',
    feature_name = 'GeneName',
    features = names(igraph::V(net)),
    column_names = c('ENSG', 'OmicsScore', 'GeneticsScore', 'Logsdon'),
    synap_import = syn$synapse
  ), 
  table_pull(
    syn_id ='syn22758536',
    feature_name = 'GName',
    features = names(igraph::V(net)),
    column_names = c('ENSG', 'GName', 'RNA_TE', 'Pro_TE'),
    synap_import = syn$synapse
  ), 
  by = 'ENSG'
 )
 colnames(omics_scores)[ colnames(omics_scores) == 'Logsdon' ] <- 'Overall'
 omics_scores <- rm_dups(
   df = omics_scores,
   feature_col = 'GName',
   feature = 'POLR2J3',
   type = 'value',
   type_spec = 'Overall' ,
   ensg_keep = NULL
 )
 omics_scores <- rm_dups(
    df = omics_scores,
    feature_col = 'GName',
    feature = "FCGBP",
    type = 'ensg',
    type_spec = 'ENSG' ,
    ensg_keep = 'ENSG00000281123'
 )
 omics_scores <- omics_scores[ !(is.na(omics_scores$GName)), ]
 row.names(omics_scores) <- omics_scores$GName
 
 net <- vertex_annotate(
   omics_scores,
   slim_net,
   "Overall",
   default_value = 0
 )   
} 
}
