% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_load}
\alias{list_load}
\title{Loads a Gene List From Synapse}
\usage{
list_load(file_path, network, is_syn = FALSE, synap_import = NULL)
}
\arguments{
\item{file_path}{the igraph network to push to synapse eg. net}

\item{network}{igraph network object to use to filter the vertices From}

\item{is_syn}{is the list a synapse ID default=FALSE}

\item{synap_import}{is the reticulated imported synapse from 
log_into_synapse()$synapse default value is NULL}
}
\value{
genes from the input present within the network
}
\description{
Loads a gene list. If is_syn is TRUE file_path is interperated as a synapse
ID. Otherwise file_path is interperated as a file path.
}
\examples{
data(slim_net)
all_goterms <- list(
   c(
       'syn25185319', 
       system.file(
           "extdata/inputlists/", 
           "APP_Metabolism.txt", 
           package = "igraphNetworkExpansion"), 
       "APP Metabolism"
   ),
  c(
       'syn25185320', 
       system.file(
           "extdata/inputlists/", 
           "Endolysosomal.txt", 
           package = "igraphNetworkExpansion"), 
       "Endolysosomal"
   )
)
list_load(
   all_goterms[[1]][2],
   network = slim_net
)
}
