% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{network_load}
\alias{network_load}
\title{Re-Load the  networks and calc meterics}
\usage{
network_load(syn_id, form)
}
\arguments{
\item{syn_id}{the networks synID}

\item{form}{the format of the netwrok file eg. "graphml"}
}
\value{
an igraph network object
}
\description{
This Function Loads a network object from synapse using its synapse ID and
the type of network, ie graphml.
}
\examples{

}
