% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphplot2.R
\name{plot.igraph2}
\alias{plot.igraph2}
\title{Modified igraph plotting code to allow for changes in edge.arrow.width.
   Edge.arrow.size still not supported
Code provided from: jevansbio/igraphhack
Git User: jevansbio}
\usage{
plot.igraph2(
  x,
  axes = FALSE,
  add = FALSE,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  mark.groups = list(),
  mark.shape = 1/2,
  mark.col = rainbow(length(mark.groups), alpha = 0.3),
  mark.border = rainbow(length(mark.groups), alpha = 1),
  mark.expand = 15,
  ...
)
}
\description{
Modified igraph plotting code to allow for changes in edge.arrow.width.
   Edge.arrow.size still not supported
Code provided from: jevansbio/igraphhack
Git User: jevansbio
}
